Refactored schema.py into package with dedicated source files for types, models, and views. Updated type annotations so that the schema package type checks with strict mode.
