+++ autograder/core/models/__init__.py
@@ -1,26 +1,45 @@
 # Import all Model classes here.
 
-from .ag_command import AGCommandResult, AGCommandResultBase, Command
-from .ag_model_base import AutograderModel
-from .ag_test.ag_test_case import AGTestCase, AGTestCaseFeedbackConfig
-from .ag_test.ag_test_case_result import AGTestCaseResult
-from .ag_test.ag_test_command import (AGTestCommand, ExpectedOutputSource, ExpectedReturnCode,
-                                      AGTestCommandFeedbackConfig, StdinSource,
-                                      ValueFeedbackLevel)
-from .ag_test.ag_test_command_result import AGTestCommandResult
-from .ag_test.ag_test_suite import AGTestSuite, AGTestSuiteFeedbackConfig
-from .ag_test.ag_test_suite_result import AGTestSuiteResult
-from .ag_test.feedback_category import FeedbackCategory
-from .course import Course, LateDaysRemaining, Semester
-from .group import Group, GroupInvitation
-from .mutation_test_suite import (BugsExposedFeedbackLevel, MutationTestSuite,
-                                  MutationTestSuiteFeedbackConfig, MutationTestSuiteResult)
-from .project import Project, UltimateSubmissionPolicy
-from .project.download_task import DownloadTask, DownloadType
-from .project.expected_student_file import ExpectedStudentFile
-from .project.instructor_file import InstructorFile
-from .rerun_submissions_task import RerunSubmissionsTask
-from .sandbox_docker_image import BuildImageStatus, BuildSandboxDockerImageTask, SandboxDockerImage
-from .submission import (Submission, get_mutation_test_suite_results_queryset,
-                         get_submissions_with_results_queryset)
-from .task import Task
+from .ag_command import AGCommandResult as AGCommandResult
+from .ag_command import AGCommandResultBase as AGCommandResultBase
+from .ag_command import Command as Command
+from .ag_model_base import AutograderModel as AutograderModel
+from .ag_test.ag_test_case import AGTestCase as AGTestCase
+from .ag_test.ag_test_case import AGTestCaseFeedbackConfig as AGTestCaseFeedbackConfig
+from .ag_test.ag_test_case_result import AGTestCaseResult as AGTestCaseResult
+from .ag_test.ag_test_command import AGTestCommand as AGTestCommand
+from .ag_test.ag_test_command import AGTestCommandFeedbackConfig as AGTestCommandFeedbackConfig
+from .ag_test.ag_test_command import ExpectedOutputSource as ExpectedOutputSource
+from .ag_test.ag_test_command import ExpectedReturnCode as ExpectedReturnCode
+from .ag_test.ag_test_command import StdinSource as StdinSource
+from .ag_test.ag_test_command import ValueFeedbackLevel as ValueFeedbackLevel
+from .ag_test.ag_test_command_result import AGTestCommandResult as AGTestCommandResult
+from .ag_test.ag_test_suite import AGTestSuite as AGTestSuite
+from .ag_test.ag_test_suite import AGTestSuiteFeedbackConfig as AGTestSuiteFeedbackConfig
+from .ag_test.ag_test_suite_result import AGTestSuiteResult as AGTestSuiteResult
+from .ag_test.feedback_category import FeedbackCategory as FeedbackCategory
+from .course import Course as Course
+from .course import LateDaysRemaining as LateDaysRemaining
+from .course import Semester as Semester
+from .group import Group as Group
+from .group import GroupInvitation as GroupInvitation
+from .mutation_test_suite import BugsExposedFeedbackLevel as BugsExposedFeedbackLevel
+from .mutation_test_suite import MutationTestSuite as MutationTestSuite
+from .mutation_test_suite import MutationTestSuiteFeedbackConfig as MutationTestSuiteFeedbackConfig
+from .mutation_test_suite import MutationTestSuiteResult as MutationTestSuiteResult
+from .project import Project as Project
+from .project import UltimateSubmissionPolicy as UltimateSubmissionPolicy
+from .project.download_task import DownloadTask as DownloadTask
+from .project.download_task import DownloadType as DownloadType
+from .project.expected_student_file import ExpectedStudentFile as ExpectedStudentFile
+from .project.instructor_file import InstructorFile as InstructorFile
+from .rerun_submissions_task import RerunSubmissionsTask as RerunSubmissionsTask
+from .sandbox_docker_image import BuildImageStatus as BuildImageStatus
+from .sandbox_docker_image import BuildSandboxDockerImageTask as BuildSandboxDockerImageTask
+from .sandbox_docker_image import SandboxDockerImage as SandboxDockerImage
+from .submission import Submission as Submission
+from .submission import \
+    get_mutation_test_suite_results_queryset as get_mutation_test_suite_results_queryset
+from .submission import \
+    get_submissions_with_results_queryset as get_submissions_with_results_queryset
+from .task import Task as Task
