+++ autograder/core/models/ag_command/__init__.py
@@ -1,3 +1,3 @@
-from .command import Command
-from .ag_command_result_base import AGCommandResultBase
-from .ag_command_result import AGCommandResult
+from .command import Command as Command
+from .ag_command_result_base import AGCommandResultBase as AGCommandResultBase
+from .ag_command_result import AGCommandResult as AGCommandResult
