+++ autograder/core/models/group/__init__.py
@@ -1,2 +1,2 @@
-from .group import Group
-from .group_invitation import GroupInvitation
+from .group import Group as Group
+from .group_invitation import GroupInvitation as GroupInvitation
