+++ autograder/core/models/project/__init__.py
@@ -1,3 +1,4 @@
-from .expected_student_file import ExpectedStudentFile
-from .instructor_file import InstructorFile
-from .project import Project, UltimateSubmissionPolicy
+from .expected_student_file import ExpectedStudentFile as ExpectedStudentFile
+from .instructor_file import InstructorFile as InstructorFile
+from .project import Project as Project
+from .project import UltimateSubmissionPolicy as UltimateSubmissionPolicy
