+++ autograder/rest_api/schema/__init__.py
@@ -1,0 +1,1 @@
+from .schema import AGSchemaGenerator as AGSchemaGenerator
