+++ autograder/utils/__init__.py
@@ -1,1 +1,6 @@
-from .misc import *
+from .misc import count_if as count_if
+from .misc import find_if as find_if
+from .misc import lock_users as lock_users
+from .misc import exclude_dict as exclude_dict
+from .misc import filter_dict as filter_dict
+from .misc import two_decimal_place_string as two_decimal_place_string
