+++ run_mypy.sh
@@ -3,4 +3,4 @@ mypy \
     autograder/core/models/course.py \
     autograder/core/tests/test_models/test_course.py \
     autograder/core/tests/test_models/test_ag_model_base.py \
-    autograder/rest_api/schema.py
+    autograder/rest_api/schema
